fix: extract gateway and grpc specific code to the appropriate server options
